increase size of qreal from float to double. We have numerical stability issues with floats, and tests indicate that there is no performance difference in using doubles over floats.
[git-p4: depot-paths = "//depot/qt/main/": change = 167405]

+++ src/core/global/qglobal.h
@@ -1799,7 +1799,7 @@ inline int qIntCast(float f) { return int(f); }
 #include <QtCore/qfixedpoint.h>
 typedef QFixedPoint qreal;
 #else
-typedef float qreal;
+typedef double qreal;
 #endif
 
 //
